Latest GUI code
Implemented Justin’s database work into my IDE.  I have made changes to
the look of the GUI to be more consistent.  Also changed a few things
to remove the duplicate functionality.

Next I need to work on getting the JComboBox to be populated with
values from the database.  Specifically size and style.