Udated GUI and Controller for error handling.  Added cancelButton and CancelNew JAR File
1.  I added code that will catch problems with incorrect data being
entered into the Customer and Catalog tabs, including load from file.
If something is wrong you get a pop up now, and the chance to correct
it. It also prevents you from entering partial entries.  There is also
some assistance with the reports.
2.  Added Cancel Button and cancelSale method for Justin.
3. Updated Jar file.
4. You could probably start trying to break or test the system now.
minus the sales portion at this time.